Change serial number field type to big integer

+++ tally_ho/apps/tally/models/result_form.py
@@ -166,7 +166,7 @@ class ResultForm(BaseModel):
                                on_delete=models.PROTECT)
     rejected_count = models.PositiveIntegerField(default=0)
     reject_reason = models.TextField(null=True, blank=True)
-    serial_number = models.PositiveIntegerField(null=True)
+    serial_number = models.BigIntegerField(null=True)
     skip_quarantine_checks = models.BooleanField(default=False)
     station_number = models.PositiveSmallIntegerField(blank=True, null=True)
     is_replacement = models.BooleanField(default=False)
